Address PR review comments: refactor git tracking to base Logger class
- Move git tracking methods from MLFlowLogger to base Logger class
- Make log_git_info a public method that needs manual invocation
- Use functools.partial for cleaner subprocess.run calls
- Replace cwd parameter with git -C flag in all git commands
- Use git diff-index for checking dirty state
- Use git branch --show-current for getting branch name
- Use git remote get-url origin for remote URL
- Remove log_storage_info method per reviewer feedback
